pyproject.toml
[tool.poetry]
name = "jupyter-grader"
version = "0.1.0"
description = ""
authors = ["Evgeny Tsydenov <evgenytsydenov@gmail.com>"]
readme = "README.md"
packages = [{include = "jupyter_grader", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
nbgrader = "^0.8.1"
google-api-python-client = "^2.70.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
python-dotenv = "^0.21.0"
email-validator = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.12.0"
ruff = "^0.0.187"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
select = [
    "E", "F", "W", "C90", "I", "D", "UP", "N", "YTT", "ANN", "S", "BLE", "FBT",
    "B", "A", "C4", "T10", "ICN", "T20", "Q", "SIM", "TID", "ARG",
    "ERA", "PGH", "PLC", "PLE", "PLR", "PLW", "RUF",
]
ignore = [
    "FBT001", "FBT002", "D100", "D105", "ANN101", "ANN102", "ANN002", "ANN003",
    "ANN401", "D203", "D204", "D213", "D215", "D400", "D404", "D406", "D407",
    "D408", "D409", "D413",
]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".hg", ".mypy_cache", ".nox",
    ".pants.d", ".ruff_cache", ".svn", ".tox", ".venv", "__pypackages__",
    "_build", "buck-out", "build", "dist", "node_modules", "venv",
]
target-version = "py310"
fix = true
ignore-init-module-imports = true
